UNST-7948
Fix double precision statements to compile project

+++ src/engines_gpl/dflowfm/packages/dflowfm_kernel/src/dflowfm_kernel/prepost/mass_balance_areas.f90
@@ -99,7 +99,7 @@ contains
       jamba = 1
       ibflag = 1
 
-      timembastart = tstart_user ! when DFM doesn't start at t=0.0??
+      timembastart = tstart_user ! when DFM doesn't start at t=0.0_dp??
       timembastarttot = timembastart
 
       flxdmp = 0.0
@@ -141,68 +141,68 @@ contains
 
       nombabnd = nomba + nopenbndsect
 
-      call realloc(mbaarea, nomba, keepExisting=.false., fill=0d0)
+      call realloc(mbaarea, nomba, keepExisting=.false., fill=0.0_dp)
 
-      call realloc(mbavolumebegin, nomba, keepExisting=.false., fill=0d0)
-      call realloc(mbavolumebegintot, nomba, keepExisting=.false., fill=0d0)
-      call realloc(mbavolumeend, nomba, keepExisting=.false., fill=0d0)
+      call realloc(mbavolumebegin, nomba, keepExisting=.false., fill=0.0_dp)
+      call realloc(mbavolumebegintot, nomba, keepExisting=.false., fill=0.0_dp)
+      call realloc(mbavolumeend, nomba, keepExisting=.false., fill=0.0_dp)
 
-      call realloc(mbaflowhor, [2, nombabnd, nombabnd], keepExisting=.false., fill=0d0)
-      call realloc(mbaflowhortot, [2, nombabnd, nombabnd], keepExisting=.false., fill=0d0)
-      call realloc(mbaflowsorsin, [2, numsrc], keepExisting=.false., fill=0d0)
-      call realloc(mbaflowsorsintot, [2, numsrc], keepExisting=.false., fill=0d0)
-      call realloc(mbaflowraineva, [2, nomba], keepExisting=.false., fill=0d0)
-      call realloc(mbaflowrainevatot, [2, nomba], keepExisting=.false., fill=0d0)
-      call realloc(mbafloweva, nomba, keepExisting=.false., fill=0d0)
-      call realloc(mbaflowevatot, nomba, keepExisting=.false., fill=0d0)
+      call realloc(mbaflowhor, [2, nombabnd, nombabnd], keepExisting=.false., fill=0.0_dp)
+      call realloc(mbaflowhortot, [2, nombabnd, nombabnd], keepExisting=.false., fill=0.0_dp)
+      call realloc(mbaflowsorsin, [2, numsrc], keepExisting=.false., fill=0.0_dp)
+      call realloc(mbaflowsorsintot, [2, numsrc], keepExisting=.false., fill=0.0_dp)
+      call realloc(mbaflowraineva, [2, nomba], keepExisting=.false., fill=0.0_dp)
+      call realloc(mbaflowrainevatot, [2, nomba], keepExisting=.false., fill=0.0_dp)
+      call realloc(mbafloweva, nomba, keepExisting=.false., fill=0.0_dp)
+      call realloc(mbaflowevatot, nomba, keepExisting=.false., fill=0.0_dp)
 
-      call realloc(mbamassbegin, [nombs, nomba], keepExisting=.false., fill=0d0)
-      call realloc(mbamassbegintot, [nombs, nomba], keepExisting=.false., fill=0d0)
-      call realloc(mbamassend, [nombs, nomba], keepExisting=.false., fill=0d0)
+      call realloc(mbamassbegin, [nombs, nomba], keepExisting=.false., fill=0.0_dp)
+      call realloc(mbamassbegintot, [nombs, nomba], keepExisting=.false., fill=0.0_dp)
+      call realloc(mbamassend, [nombs, nomba], keepExisting=.false., fill=0.0_dp)
 
       if (stm_included) then
-         call realloc(mbabedmassbegin, [lsedtot, nomba], keepExisting=.false., fill=0d0)
-         call realloc(mbabedmassbegintot, [lsedtot, nomba], keepExisting=.false., fill=0d0)
-         call realloc(mbabedmassend, [lsedtot, nomba], keepExisting=.false., fill=0d0)
+         call realloc(mbabedmassbegin, [lsedtot, nomba], keepExisting=.false., fill=0.0_dp)
+         call realloc(mbabedmassbegintot, [lsedtot, nomba], keepExisting=.false., fill=0.0_dp)
+         call realloc(mbabedmassend, [lsedtot, nomba], keepExisting=.false., fill=0.0_dp)
 
-         call realloc(mbabedshortmassbegin, [lsedtot, nomba], keepExisting=.false., fill=0d0)
-         call realloc(mbabedshortmassbegintot, [lsedtot, nomba], keepExisting=.false., fill=0d0)
-         call realloc(mbabedshortmassend, [lsedtot, nomba], keepExisting=.false., fill=0d0)
+         call realloc(mbabedshortmassbegin, [lsedtot, nomba], keepExisting=.false., fill=0.0_dp)
+         call realloc(mbabedshortmassbegintot, [lsedtot, nomba], keepExisting=.false., fill=0.0_dp)
+         call realloc(mbabedshortmassend, [lsedtot, nomba], keepExisting=.false., fill=0.0_dp)
 
          if (iflufflyr > 0) then
-            call realloc(mbafluffmassbegin, [lsedtot, nomba], keepExisting=.false., fill=0d0)
-            call realloc(mbafluffmassbegintot, [lsedtot, nomba], keepExisting=.false., fill=0d0)
-            call realloc(mbafluffmassend, [lsedtot, nomba], keepExisting=.false., fill=0d0)
+            call realloc(mbafluffmassbegin, [lsedtot, nomba], keepExisting=.false., fill=0.0_dp)
+            call realloc(mbafluffmassbegintot, [lsedtot, nomba], keepExisting=.false., fill=0.0_dp)
+            call realloc(mbafluffmassend, [lsedtot, nomba], keepExisting=.false., fill=0.0_dp)
          end if
 
-         call realloc(mbasedflux, [2, lsedtot, nombabnd, nombabnd], keepExisting=.false., fill=0d0)
-         call realloc(mbasedfluxtot, [2, lsedtot, nombabnd, nombabnd], keepExisting=.false., fill=0d0)
+         call realloc(mbasedflux, [2, lsedtot, nombabnd, nombabnd], keepExisting=.false., fill=0.0_dp)
+         call realloc(mbasedfluxtot, [2, lsedtot, nombabnd, nombabnd], keepExisting=.false., fill=0.0_dp)
          if (jampi == 1) then
-            call realloc(mbasedfluxreduce, [2, lsedtot, nombabnd, nombabnd], keepExisting=.false., fill=0d0)
+            call realloc(mbasedfluxreduce, [2, lsedtot, nombabnd, nombabnd], keepExisting=.false., fill=0.0_dp)
          end if
       end if
 
-      call realloc(mbafluxhor, [2, numconst, nombabnd, nombabnd], keepExisting=.false., fill=0d0)
-      call realloc(mbafluxhortot, [2, numconst, nombabnd, nombabnd], keepExisting=.false., fill=0d0)
-      call realloc(mbafluxsorsin, [2, 2, numconst, numsrc], keepExisting=.false., fill=0d0)
-      call realloc(mbafluxsorsintot, [2, 2, numconst, numsrc], keepExisting=.false., fill=0d0)
-      call realloc(mbafluxheat, [2, nomba], keepExisting=.false., fill=0d0)
-      call realloc(mbafluxheattot, [2, nomba], keepExisting=.false., fill=0d0)
+      call realloc(mbafluxhor, [2, numconst, nombabnd, nombabnd], keepExisting=.false., fill=0.0_dp)
+      call realloc(mbafluxhortot, [2, numconst, nombabnd, nombabnd], keepExisting=.false., fill=0.0_dp)
+      call realloc(mbafluxsorsin, [2, 2, numconst, numsrc], keepExisting=.false., fill=0.0_dp)
+      call realloc(mbafluxsorsintot, [2, 2, numconst, numsrc], keepExisting=.false., fill=0.0_dp)
+      call realloc(mbafluxheat, [2, nomba], keepExisting=.false., fill=0.0_dp)
+      call realloc(mbafluxheattot, [2, nomba], keepExisting=.false., fill=0.0_dp)
 
       if (.not. allocated(srcname)) then
          allocate (srcname(0))
       end if
 
       if (jampi == 1) then
-         call realloc(mbavolumereduce, nomba, keepExisting=.false., fill=0d0)
-         call realloc(mbaflowhorreduce, [2, nombabnd, nombabnd], keepExisting=.false., fill=0d0)
-         call realloc(mbaflowsorsinreduce, [2, numsrc], keepExisting=.false., fill=0d0)
-         call realloc(mbaflowrainevareduce, [2, nomba], keepExisting=.false., fill=0d0)
-         call realloc(mbaflowevareduce, nomba, keepExisting=.false., fill=0d0)
-         call realloc(mbamassreduce, [nombs, nomba], keepExisting=.false., fill=0d0)
-         call realloc(mbafluxhorreduce, [2, numconst, nombabnd, nombabnd], keepExisting=.false., fill=0d0)
-         call realloc(mbafluxsorsinreduce, [2, 2, numconst, numsrc], keepExisting=.false., fill=0d0)
-         call realloc(mbafluxheatreduce, [2, nomba], keepExisting=.false., fill=0d0)
+         call realloc(mbavolumereduce, nomba, keepExisting=.false., fill=0.0_dp)
+         call realloc(mbaflowhorreduce, [2, nombabnd, nombabnd], keepExisting=.false., fill=0.0_dp)
+         call realloc(mbaflowsorsinreduce, [2, numsrc], keepExisting=.false., fill=0.0_dp)
+         call realloc(mbaflowrainevareduce, [2, nomba], keepExisting=.false., fill=0.0_dp)
+         call realloc(mbaflowevareduce, nomba, keepExisting=.false., fill=0.0_dp)
+         call realloc(mbamassreduce, [nombs, nomba], keepExisting=.false., fill=0.0_dp)
+         call realloc(mbafluxhorreduce, [2, numconst, nombabnd, nombabnd], keepExisting=.false., fill=0.0_dp)
+         call realloc(mbafluxsorsinreduce, [2, 2, numconst, numsrc], keepExisting=.false., fill=0.0_dp)
+         call realloc(mbafluxheatreduce, [2, nomba], keepExisting=.false., fill=0.0_dp)
       end if
 
 !  Determine 2D pointers fo links (from balance area to balance area)
@@ -530,7 +530,7 @@ contains
       use m_flowtimes, only: refdate_mjd
       use time_module, only: mjd2date
 
-      double precision, intent(in) :: time !< time represented in seconds since refdate
+      real(dp), intent(in) :: time !< time represented in seconds since refdate
       character(len=19), intent(out) :: datestr !< time represented as string
 
       integer :: iyear !< year
@@ -538,7 +538,7 @@ contains
       integer :: iday !< day number
       integer :: ihour !< hours
       integer :: imin !< minutes
-      double precision :: sec !< (fractional) seconds
+      real(dp) :: sec !< (fractional) seconds
 
       datestr = ""
       if (mjd2date(refdate_mjd + time / 86400.0_dp, iyear, imonth, iday, ihour, imin, sec) /= 0) then
